Refactor Playlist and Controls into single Player
- Move refresh into player
- Compose Player into handlers
- Try to encapsulate MPD stuff in MpdPlayer
